Create process reading to true or false
This node processes the readings of the ultrasonic sensor and determines whether there was activity in the room.
We define activity, after some testing, to be and change in ultrasonic readings by more than 20 cm. Hence, in order to 'detect activity', we write the codes to process all the readings. Since our project requires room activity every 30 mins and the readings are done every 5 mins (as per the previous node), we create an array of readings and store it up to 6 times before processing if there was activity. 

Firstly (1-4) creates a variable 'intervals' if it does not exists as a global variable when u run the code for the first time. 
We will add the ultrasonic sensor readings into the array that the previous node gave us.
While the array is still below 6 readings, the code will continue at line 25.
(25) This is important, as we have to store the values in the intervals, otherwise the intervals will always have only 1 readings as 'intervals' is undefined.
One thing to take note when using global variables, as there might be instances where the global variable has too many readings stored when testing the codes (as its all in a global variable) and cause it to not enter the while loop below. In that case, we need to set intervals = [] to clear the cache before it will work again.

 (7) This will continue until the array reaches 6 readings, meaning the 30 min mark has reached.
(8-10) To determine activity, we will find out the range between the smallest and biggest readings in the array
(12-17) If the range is more than 20 cm, we define it as there having activity in the room and sets the msg.payload to a json format with the sensor id as well as status being true. This will then be sent in the next node via API requests to the database.
(18-23) Similar to previously, but for a false message instead.